# project-management-system/PMS-Swagger- Documentation.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SriramB/ProjectManagementSystem/1.0.0
  - description: URL for PMS
    url: http://localhost:8080/api/v1
info:
  description: This API is PMS
  version: "1.0.0"
  title: PMS
  contact:
    email: suba@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /service/getServices:
    get:
      tags:
        - developers
        - analyst
        - manager
        - tester
        - admins
      security:
        - bearerAuth: []
      summary: Search for All Services
      operationId: getServices
      description: |
        Search for All Services
      responses:
        '200':
          description: List of Services
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ServicesList'
        '400':
          description: Invalid Input
  /service/getServiceByName/{serviceName}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for particular Services
      operationId: getServiceByName
      description: |
        Search for particular Service
      parameters:
        - in: path
          name: serviceName
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: One Services
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Service'
        '400':
          description: Invalid Input
  /service/getServiceByRole/{role}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
        - bearerAuth: []
      summary: Search for particular Services
      operationId: getServiceByRole
      description: |
        Search for particular Service
      parameters:
        - in: path
          name: role 
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: One Services
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/RoleService'
        '400':
          description: Invalid Input
  /service/updatePermission/{serviceName}:
    patch:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: update existing Role Permissions
      operationId: updateService
      description: |
        update existing Role Permissions
      parameters:
        - in: path
          name: serviceName
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/updateService'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateListContent'
  /service/removePermission/{serviceName}:
    patch:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: Removing existing Role Permissions
      operationId: removePermission
      description: |
       Removing existing Role Permissions
      parameters:
        - in: path
          name: serviceName
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/updateService'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateListContent'
  /service/addService:
    post:
      tags:
        - admins
        - developers
        - manager
      security:
      - bearerAuth: []
      summary: Add a new  Service
      operationId: addService
      description: |
        Add a new Service
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/InsertedService'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceListContent'
  /service/deleteService/{serviceName}:
    delete:
      tags:
        - admins
        - manager
      security:
      - bearerAuth: []
      summary: delete an existing Service
      operationId: deleteService
      description: |
        delete an existing Service
      parameters:
        - in: path
          name: serviceName
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedService'
        '400':
          description: Invalid Input
...
  /TestcaseService/addTestcase:
    post:
      tags:
        - admins
        - tester
        - developers
      security:
      - bearerAuth: []
      summary: Add a new Testcase
      operationId: addTestcase
      description: |
        Adding a new Testcase
      responses:
        '200':
          description: Testcase Successfully Added
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/AddedTestcase'
        '400':
          description: Invalid Input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestcaseListContent'
  /TestcaseService/getTestcases:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
      - bearerAuth: []
      summary: Getting All Testcases 
      operationId: getTestcase
      description: |
        Getting a specific Testcase
      responses:
        '200':
          description: All Testcases fetched Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Testcases'
        '400':
          description: Invalid Input
  /TestcaseService/getTestcasesByRid/{requirementId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
      - bearerAuth: []
      summary: Getting All Testcases By Requirement ID
      operationId: getTestcaseByRid
      parameters:
        - in: path
          name: requirementId
          required: true
          schema: 
            type: string
      description: |
        Getting a specific Testcase
      responses:
        '200':
          description: All Testcases fetched Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Testcases'
        '400':
          description: Invalid Input
  /TestcaseService/getTestcasesByPid/{projectId}:
    get:
      tags:
        - admins
        - developers
        - analyst
        - manager
        - tester
      security:
      - bearerAuth: []
      summary: Getting All Testcases By Project ID
      operationId: getTestcaseByPid
      parameters:
        - in: path
          name: projectId
          required: true
          schema: 
            type: string
      description: |
        Getting a specific Testcase
      responses:
        '200':
          description: All Testcases fetched Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Testcases'
        '400':
          description: Invalid Input
  /TestcaseService/editTestcase/{testcaseId}:
    patch:
      tags:
        - admins
        - tester
        - developers
      security:
      - bearerAuth: []
      summary: update an existing Testcase
      operationId: updateTestcase
      description: |
        update an existing Project Requirement
      parameters:
        - in: path
          name: testcaseId
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestcaseEditContent'
      responses:
        '200':
          description: Edited Requirement Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/EditedTestcase'
        '400':
          description: Invalid Input
  /TestcaseService/deleteTestcase/{testcaseId}:
    delete:
      tags:
        - admins
        - tester
      security:
      - bearerAuth: []
      summary: delete an existing Testcase
      operationId: deleteTestcase
      description: |
        update an existing Project Requirement
      parameters:
        - in: path
          name: testcaseId
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: Deleted Testcase Successfully
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/DeletedTestcase'
        '400':
          description: Invalid Input            
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServicesList:
      type: object
      required:
        - Services
      properties:
        Services:
          items:
              $ref: '#/components/schemas/ServiceListContent'
    ServiceDeletedListContent:
      type: object
      required:
        - deletedCount
      properties:
        deletedCount:
          type: integer
          example: 1
    ServiceListContent:
      type: object
      required:
        - serviceName
      properties:
        serviceName:
          type: string
          example: addRole
        rolesPermission:
          type: object
          example: ["developer","tester"]
        createdBy :
          type: string
          example: "U14"
        LastUpdatedBy:
          type: string
          example: "U15"
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
    ServiceUpdateListContent:
      type: object
      required:
        - rolesPermission
      properties:
        rolesPermission:
          type: object
          example: ["developer","tester"]
    Service:
      type: object
      required: 
        - Service
      properties:
        ServiceDetail:
          $ref: '#/components/schemas/ServiceListContent'     
    InsertedService:
      type: object
      required: 
        - Service
      properties:
        Inserted:
          $ref: '#/components/schemas/ServiceListContent'
    DeletedService:
      type: object
      required: 
        - Service
      properties:
        Deleted:
          $ref: '#/components/schemas/ServiceDeletedListContent'
    RoleService:
      type: object
      required: 
        - Service
      properties:
        Services:
          $ref: '#/components/schemas/ServiceListContent'      
...
    updateService:
      type: object
      required:
        - Service
      properties:
        Updated:
          items:
              $ref: '#/components/schemas/ServiceListContent'
...
    TestcaseEditContent:
     type: object
     properties:
        testcaseName:
          type: string
          example: Testing Login screen for chat
        testcaseDescription:
          type: string
          example: A short msg should appear after login
        expectedResults:
          type: string
          example: Login Succesful in Toast
        actualResults:
          type: string
          example: XXX
...
    TestcaseListContent:
      type: object
      properties:
        testcaseId:
          type: string
          example: T1
        requirementId:
          type: string
          example: P1R1
        projectId:
          type: string
          example: P1
        testcaseName:
          type: string
          example: Testing Login screen for chat
        testcaseDescription:
          type: string
          example: A short msg should appear after login
        expectedResults:
          type: string
          example: Login Succesful in Toast
        actualResults:
          type: string
          example: XXX
        status:
          type: string
          example: Active
        createdBy :
          type: string
          example: U14
        UpdatedBy:
          type: string
          example: U15
        LastUpdatedDate:
          type: object
          example: "2020-11-10T14:45:45.666+00:00"
...
    AddedTestcase:
      type: object
      required: 
        - Role
      properties:
        Added Testcase:
          $ref: '#/components/schemas/TestcaseListContent'
    Testcases:
      type: object
      required:
        - Roles
      properties:
        Testcases:
          items:
              $ref: '#/components/schemas/TestcaseListContent'
    EditedTestcase:
      type: object
      required:
        - Roles
      properties:
        Edited Testcase:
          items:
              $ref: '#/components/schemas/TestcaseListContent'
    DeletedTestcase:
      type: object
      required:
        - Roles
      properties:
        Deleted Testcase:
          items:
              $ref: '#/components/schemas/TestcaseListContent'